Added styled-components and styled-reset Cleaned up and organized directory structure
