+++ web-workout-tracker/workout/src/design/global_style.js
@@ -1,0 +1,24 @@
+import {createGlobalStyle} from 'styled-components'
+import {reset} from 'styled-reset'
+
+const GlobalStyle = createGlobalStyle`
+    ${reset}
+    //FONTS
+    //STYLE
+    * {box-sizing: border-box}
+    html {
+        font-size: 62.5%;
+        height: 100%;
+        width: 100%;
+    }
+    body {
+        height: 100%;
+        width: 100%;
+    }
+    a {
+        color: inherit;
+        text-decoration: none;
+    }
+`
+
+export default GlobalStyle
