+++ web-workout-tracker/workout/src/views/note.text
@@ -1,0 +1,1 @@
+components tied to the router / NOT REUSABLE
